frameworks/files: Naive script-only files.log unrolling
This is a script-only change that unrolls File::Info records into
multiple files.log entries if the file was seen over multiple
connections. This makes tx_hosts, rx_hosts and conn_uids as sets
not meaningful and so these are removed by default from the log
and also from the File::Info record.

To get back the fields in the log and in script land,
ZEEK_FILES_LOG_INCLUDE_SETS=1 can be set as an environment variable
to bring them back. Unrolling is not configurable in this version.
The motivation being that in a Zeek cluster where multiple workers
see different parts of file-transfers, the files.log is already
partially unrolled.

Test commit to see where this could be going.