Truncate Docker image artifact at end of Docker CI workflow
In the (private) zeek-security repo we currently hit the 2GB storage limit
whenever CI runs happen multiple times in one day, because our Docker Github
workflow preserves the output of the first job (the image build) for the second
job (the cluster tests).

For successful runs, there's no need for the Docker image to survive the CI
run. (For failures, it might help to grab the image for local testing.)

There's no Github-official way to delete an attachment within a workflow, so
this simply truncates the artifact if the workflow succeeds.