zeekygen/normalize_script_path: Special case plugin dirnames without _
In normal installations, a plugin's basename has an underscore in it
to separate the namespace from the plugin name. E.g Zeek_Spicy. When
there is no underscore, this is most likely due to ./build being
picked up when using ZEEK_PLUGIN_PATH. The basename ends-up "build"
and is susceptible to collisions.

Prepend one parent directory as a heuristic to make this scenario less
likely, assuming ./build is usually below a repository checkout that
uniquely identifies the plugin.

Fixes #2577