ssl: Make disable_analyzer_after_detection independent from logging
Basically, move analyzer_id out of the Info record and directly onto
connection. Further, do not require set_session() for finish() to
function.

This break consumers that relied on ssl$analyzer_id's existence and
requires them to move to c$ssl_analyzer_id instead. We could duplicate
the ID onto the Info field for backwards compat, but not sure that's
worth it.