[RFC] dump-events: Move into base and use event groups for toggling
Mostly to measure the perf difference for a heavy hitter script and
generally ignite the thought of replacing @load functionality available
at runtime or via const/option values, assuming the perf impact is
negligible.

For dump-events, it's unclear if someone would turn it on at runtime,
but possibly other scripts could be better examples.