cirrus: Add test for builtin plugins
This is very smokey and adds a dependency on external projects which
may hurt us if they break something. Once that happens, we can either
remove them or decide if we want to pin to a certain commit. At this
point I'm somewhat liking this as an early integration. We previously
had a user report that bro-http2 failed to build and zeek-kafka seems
popular enough.

Unfortunately zeek/zeek#2806 is affecting this exercise.

Closes #2837